housesRoutes: added jwt auth in POST, PATCH and DELETE routes

+++ routes/housesRoutes.js
@@ -1,28 +1,40 @@
 // Import files
 import { Router } from 'express'
 import db from '../db.js'
+import { jwtSecret } from '../secrets.js'
+import jwt from 'jsonwebtoken'
 
 // Create router
 const router = Router()
 
 // POST request route for creating houses
 router.post('/houses', async (req, res) => {
-  const {
-    location,
-    bedrooms,
-    bathrooms,
-    price_per_night,
-    description,
-    host_id
-  } = req.body
+  // getting the jwt token from cookies
+  let token = req.cookies.jwt
 
+  // deconstructing user_id from the jwt token in cookies
+  const { user_id } = jwt.verify(token, jwtSecret)
+
+  // deconstructing req.body to get listing data
+  const { location, bedrooms, bathrooms, price_per_night, description } =
+    req.body
+
+  // final query that will be sent to db
   const finalQuery = `INSERT INTO houses (location, bedrooms, bathrooms, price_per_night, description, host_id)
-  VALUES ('${location}', ${bedrooms}, ${bathrooms}, ${price_per_night}, '${description}', ${host_id})
+  VALUES ('${location}', ${bedrooms}, ${bathrooms}, ${price_per_night}, '${description}', ${user_id})
   RETURNING *
   `
 
   try {
+    // if user_id is not found, or token does not exist or is not valid, throw error
+    if (!user_id || !token) {
+      res.json({ error: 'Invalid authentication token' })
+    }
+
     const { rows } = await db.query(finalQuery)
+
+    if (!rows.length)
+      throw new Error('Error creating the listing, please try again.')
 
     res.json(rows)
   } catch (err) {
@@ -104,6 +116,12 @@ router.get('/houses/:house_id', async (req, res) => {
 
 // PATCH ROUTE - Update a house based on Payload from PATCH request
 router.patch('/houses/:house_id', async (req, res) => {
+  // getting the jwt token from cookies
+  let token = req.cookies.jwt
+
+  // deconstructing user_id from the jwt token in cookies
+  const { user_id } = jwt.verify(token, jwtSecret)
+
   // Deconstruct body of the PATCH request
   const { location, bedrooms, bathrooms, price_per_night, description } =
     req.body
@@ -143,6 +161,22 @@ router.patch('/houses/:house_id', async (req, res) => {
 
   // UPDATE Database
   try {
+    // if user_id is not found, or token does not exist or is not valid, throw error
+    if (!user_id || !token) {
+      res.json({ error: 'Invalid authentication token' })
+    }
+
+    // authorization check to see if host_id matches the user_id from jwt token
+    const houseObj = await db.query(
+      `SELECT * FROM houses WHERE house_id = ${house_id}`
+    )
+
+    // get host_id from response
+    const host_id = houseObj.rows[0].host_id
+
+    // throw new error if user_id does not match host_id from db
+    if (host_id !== user_id) throw new Error('You are not authorized')
+
     const { rows } = await db.query(finalQueryStr)
     if (!rows.length) {
       throw new Error('There was an error with the update')
@@ -156,8 +190,30 @@ router.patch('/houses/:house_id', async (req, res) => {
 // DELETE ROUTE
 // Define a route for deleting individual house
 router.delete('/houses/:house_id', async (req, res) => {
+  // getting the jwt token from cookies
+  let token = req.cookies.jwt
+
+  // deconstructing user_id from the jwt token in cookies
+  const { user_id } = jwt.verify(token, jwtSecret)
+
   let house_id = req.params.house_id
   try {
+    // if user_id is not found, or token does not exist or is not valid, throw error
+    if (!user_id || !token) {
+      res.json({ error: 'Invalid authentication token' })
+    }
+
+    // authorization check to see if host_id matches the user_id from jwt token
+    const houseObj = await db.query(
+      `SELECT * FROM houses WHERE house_id = ${house_id}`
+    )
+
+    // get host_id from response
+    const host_id = houseObj.rows[0].host_id
+
+    // throw new error if user_id does not match host_id from db
+    if (host_id !== user_id) throw new Error('You are not authorized')
+
     const { rows } = await db.query(
       `DELETE FROM houses WHERE house_id = ${house_id} RETURNING *`
     )
